[HUDI-4052] Move HoodieMetricsCloudWatchConfig to hudi-client-common
